refactor(cuaktask): move isPromiseLike function to common
