+++ packages/cuaktask/src/common/utils/domain/isPromiseLike.spec.ts
