+++ packages/cuaktask/src/common/utils/domain/isPromiseLike.ts
