+++ packages/cuaktask/src/index.ts
@@ -1,3 +1,4 @@
+import { isPromiseLike } from './common/utils/domain/isPromiseLike';
 import { BaseDependentTask } from './task/models/domain/BaseDependentTask';
 import { BaseTask } from './task/models/domain/BaseTask';
 import { DependentTask } from './task/models/domain/DependentTask';
@@ -8,7 +9,6 @@ import { TaskStatus } from './task/models/domain/TaskStatus';
 import { DependentTaskBuilder } from './task/modules/DependentTaskBuilder';
 import { DependentTaskRunner } from './task/modules/DependentTaskRunner';
 import { TaskDependencyEngine } from './task/modules/TaskDependencyEngine';
-import { isPromiseLike } from './utils/isPromiseLike';
 
 export type {
   DependentTask,
